refactor: modularize optimal cutoffs
